Network.HTTP.Cookie

Description

This module provides the data types and functions for working with HTTP cookies. Right now, it contains mostly functionality needed by Browser.

Documentation

data Cookie

Cookie is the Haskell representation of HTTP cookie values. See its relevant specs for authoritative details.

Constructors

MkCookie

Fields

ckDomain :: String

ckName :: String

ckValue :: String

ckPath :: Maybe String

ckComment :: Maybe String

ckVersion :: Maybe String

cookieMatch :: (String, String) -> Cookie -> Bool

cookieMatch (domain,path) ck performs the standard cookie match wrt the given domain and path.

cookiesToHeader :: [Cookie] -> Header

cookieToHeaders ck serialises Cookies to an HTTP request header.

processCookieHeaders :: String -> [Header] -> ([String], [Cookie])

processCookieHeaders dom hdrs